+++ packages/openapi_repository_annotations/CHANGELOG.md
@@ -1,3 +1,7 @@
-## 1.0.0
+## [1.1.0] - 13 Jan 2023
+
+- Implemented providing List of `BlocMixin` to allow adding `mixin` on ListBloc/DataBloc for `@OpenapiRepository` annotation.
+
+## [1.0.0] - 17 Aug 2022
 
 - Initial version. Contains `@OpenapiRepository` annotation that can be used with openapi_repository package